+++ src/ui/react/settings/state/mod.ts
@@ -14,7 +14,7 @@ subscribe(SeelenEvent.StateSettingsChanged, ({ payload }) => {
   initialSettings.value = JSON.stringify(payload);
 });
 
-export const language = computed(() => settings.value.language || "en");
+export const language = computed(() => settings.value.language);
 
 export const hasChanges = computed(() => initialSettings.value !== JSON.stringify(settings.value));
 export const needRestart = signal(false);