+++ src/ui/svelte/keyboard-selector/state.svelte.ts
@@ -7,7 +7,7 @@ subscribe(SeelenEvent.StateSettingsChanged, settings.setByPayload);
 await settings.init();
 $effect.root(() => {
   $effect(() => {
-    locale.set(settings.value.language || "en");
+    locale.set(settings.value.language);
   });
 });