+++ src/ui/svelte/power-menu/state.svelte.ts
@@ -8,7 +8,7 @@ await settings.init();
 
 $effect.root(() => {
   $effect(() => {
-    locale.set(settings.value.language || "en");
+    locale.set(settings.value.language);
   });
 });