+++ src/ui/svelte/quick-settings/state.svelte.ts
@@ -9,7 +9,7 @@ await settings.init();
 
 $effect.root(() => {
   $effect(() => {
-    locale.set(settings.value.language || "en");
+    locale.set(settings.value.language);
   });
 });