+++ src/ui/svelte/user-menu/state/mod.svelte.ts
@@ -11,7 +11,7 @@ await settings.init();
 
 $effect.root(() => {
   $effect(() => {
-    locale.set(settings.value.language || "en");
+    locale.set(settings.value.language);
   });
 });